rename aux.py to preprocessor.py to fix windows issue
